+++ attributes/default.rb
@@ -1,11 +1,10 @@
 default['logstash']['basedir'] = '/opt/logstash'
 default['logstash']['user'] = 'logstash'
 default['logstash']['group'] = 'logstash'
-default['logstash']['join_groups'] = []
-default['logstash']['log_dir'] = '/var/log/logstash'
+default['logstash']['supervisor_gid'] = '' # set this if you want all logstash processes to run as different gid,  can be overridden
 default['logstash']['pid_dir'] = '/var/run/logstash'
 default['logstash']['create_account'] = true
-
+default['logstash']['join_groups'] = []
 # roles/flags for various search/discovery
 default['logstash']['graphite_role'] = 'graphite_server'
 default['logstash']['graphite_query'] = "roles:#{node['logstash']['graphite_role']} AND chef_environment:#{node.chef_environment}"