+++ attributes/server.rb
@@ -1,4 +1,5 @@
 default['logstash']['server']['version'] = '1.1.13'
+default['logstash']['server']['log_file'] = '/var/log/logstash/server.log'
 default['logstash']['server']['source_url'] = 'https://logstash.objects.dreamhost.com/release/logstash-1.1.13-flatjar.jar'
 default['logstash']['server']['checksum'] = '5ba0639ff4da064c2a4f6a04bd7006b1997a6573859d3691e210b6855e1e47f1'
 default['logstash']['server']['install_method'] = 'jar' # Either `source` or `jar`
@@ -12,7 +13,7 @@ default['logstash']['server']['gc_opts'] = '-XX:+UseParallelOldGC'
 default['logstash']['server']['ipv4_only'] = false
 default['logstash']['server']['debug'] = false
 default['logstash']['server']['home'] = '/opt/logstash/server'
-default['logstash']['server']['install_rabbitmq'] = true
+default['logstash']['server']['install_rabbitmq'] = false
 
 default['logstash']['server']['init_method'] = 'native' # native or runit
 # roles/flags for various autoconfig/discovery components